vsync: synchronize vsyncs without time synchronization
This change allows us to synchronize vsyncs even if the wall clocks of
the two systems aren't synchronized. We do this by calculating the
average of its last N vsyncs that the primary has provided to us and
assuming that its next vsync would be happening at this cadence.
Then we modulo the delta with this average to give us a time difference
between it's nearest vsync to the secondary's.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>