vsync: Specify clearly that time_diff is in ms in synchronize_vsync
It wasn't clear earlier whether synchronize_vsync took time difference
in us or ms. Made it clear that this is in ms. Also, changed the formula
for calculating steps that we have to take in order to reach vsync
parity with the other system slightly to make it easier to understand.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>