vsync: Only alter all the DKL phys
Previously, we were assuming that there was only one DKL phy and it was
the 3rd one. Now, we are querying the DKL phy DIV0 register for all
possible DKL phys (up to a max of 6) and programming all of them. We
wait for all of them to be synchronized and once they are, then we
program each one to its default register set.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>