vsync: Fix the way that we capture vsync timestamps
Firstly, the vsync timestamps can't be trusted if captured by this
library. Instead, a much more precise way is to get the timestamps
from the driver. Fortunately, the DRM already provides the timestamps in
the vblank_handler as parmaters (vsec and v_usec). Also, removed an
extra relative timestamp that we were capturing.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>