vsync_test: Capture multiple vsyncs for secondary as well
Previously, we were only capturing one vsync for the secondary system.
That is not enough to figure out the average at which the vsyncs are
happening on the secondary system and we must find that average so that
we can figure out how much to speed up or slow down by and synchronize
the two systems.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>