vsync_test: Added a synchronize command line parameter for secondary
When running this program on the secondary system, we can now provide 3
parameters, ex: vsync_test sec <IP_ADDR> or a new 4th parameter called
synchronize which can be either 0 or 1. This parameter lets the program
know if we should synchronize the secondary's vsync to match with
primary's or not.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>