ptp_connection: Added a ptp_connection class
Finally, we add a PTP way to communicate between client and server. The
main benefit of using PTP over TCP is that it is much faster. Once we
have sync'd the clocks of the client/server, a message transfer is
usually done within 1 ms as opposed to ~200 ms that is sometimes seen
with TCP. PTP way does require that we either have a direct ethernet
cable between the server and client or we have PTP switches in the
middle.
You also need to provide the server's PTP interface to both the server
and client in addition to the server's PTP interface's ethernet address
also being provided to the client so that it can communicate with the
server.
For example:
On primary, run:
./vsync_test pri enp176s0

On secondary, run:
./vsync_test sec enp176s0 84:47:09:04:eb:0e

Note that the interface and address above are just examples. Replace
them with your own PTP interface name and its corresponding address.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>