message: Added vblank_count set/get functions.
Sometimes, we don't need to collect a full list of vblanks. Instead, we
are only looking for a couple of them to quickly find out how much ahead
or behind the secondary system is from primary. This change allows us to
store and retrieve the vblank counts in the message class.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>