vsyncalter: Delete the timer when we are done.
Previously, we weren't deleting this timer, and if there was a second
call made to the synchronize function, then it would've created a second
timer. Since the first one wasn't gone, both timers would fire and since
we were deleting a dynamically allocated variable, and destroying it in
the timer, the second timer would cause a seg fault as it would try to
free memory that was already freed by the first one. With this change,
we are now deleting the first timer as soon as we are done
synchronizing.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>