vsync_test: Added resync functionality if drift goes above a threshold.
Previously, we just either tested to see how far apart were the vsyncs
of the primary and secondary systems or we synch'ed the secondary once.
However, since there are two differnt reference clocks (one per system),
the two systems are invariably going to drift apart from each other. So
with this change, we have now added the capability to resync if the
secondary system goes above a certain threshold.
This change also required the command line parameters of the secondary
system to change a little. The new command line functionality works with
PTP mode like this:
./vsync_test sec enp176s0 84:47:09:04:eb:0e 100

The last parameter in the above call used to be either 0 or 1. Now, we
can add the threshold in us. So for example, in the above case, we want
to resync the two systems if they have gone above 100 us threshold. The
secondary will keep asking the primary for its last 2 timestamps, and
then compare the second one from the primary aginst its first vsync
timestamp. If it is above the 100 us threshold, then we will resync. If
it is below the 100 us threshold, then a message will simply appear
showing the exact delta between the two systems' vsyncs.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>