Combo: Added logic to find the enabled combo phys
Previously, we didn't really know what combo phys were turned on. With
this change, we are now able to detect the right combo phys and we are
also able to tune the appropriate DPLLs. There was also a bug in the
CFGCR1 register that we were using for DPLL0. Finally, there was another
bug in finding the right p and k values which we have fixed in this
commit.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>