lib: Moved to a new vtotal way to sync vsyncs
Earlier, we were synchronizing the vsyncs of two systems by modifying
the PLL registers and this was complicated because it had to change on a
per PHY basis. We had support for DKL and Combo PHYs but not for the
newer C10 and C20 synopsis PHYs found on platforms like MTL and DG2.
However, there is a much simpler mechanism to synchronize vsyncs and
that is through the vtotal/vblank modifications. The number of registers
that we have to modify are far fewer (only 2) and there is no
complicated math that we must do to get the synchronization done.
Finally, this method supports far more platforms including MTL and DG2.
So with this change, we are moving to vtotal way of synchronizing
vsyncs. We have removed all files for DKL and Combo phys and
significantly changed vsyncalter.cpp to use the vtotal sync method.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>