lib: Get Vsync interrupts on a per pipe basis
Earlier, we only used to get the vsyncs interrupts for the primary pipe.
This means, we couldn't synchronize other pipes because we didn't know
when their vsync interrupts were happening. With this change, we have
changed the get_vsync function to also take an additional pipe parameter
(that defaults to 0) so that we can get vsyncs on a per pipe basis. This
can be used in the future to synchronize other pipes as well.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>