DKL: Fix a bug with multiple DKL phys not being seen
Each Dekel PHY is addressed through a 4KB aperture. Each PHY has more than
4KB of register space, so a separate index is programmed in HIP_INDEX_REG0
or HIP_INDEX_REG1, based on the port number, to set the upper 2 address
bits that point the 4KB window into the full PHY register space.
So, basically, if we read the value of the PLL DIV0 register and it is 0,
then that means that this aperture may need to be shifted. Once we do that,
we should re-read the value of the DIV0 register once more and this time it
should be valid.
During resetting of the DKL phys, it is possible that the register index
moves back to another window and as such programming of the phy
registers has no impact. So before each programming, ensure that we are
in the right window.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>