vbltest: Provide a cmd param to get Vsync interrupts on a per pipe basis
Earlier, we only used to get the vsyncs interrupts for the primary pipe.
This means, we couldn't synchronize other pipes because we didn't know
when their vsync interrupts were happening. With this change, this
program takes an extra parameter in the form of a 0 based pipe for which
we need to get the vsyncs. So the usage of vbltest is:

vbltest <number of vsyncs to get timestamp for> <0 based pipe to get for>

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>