Combo: Fix a divide by 0 error
According to the Bspec, we have to get the qdiv_mode first from cfgcr1
register's bit 9 and only if it is enabled, should we read qdiv_ratio
from cfgcr1's bit 17:10. Otherwise, it should be 1. We were incorrectly
not reading this bit 9 and that would cause us to read bits 17:10 even
when they were 0 which would end up causing a divide by 0 error. This
change addresses this problem.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>