combo: Moved cleanup_list to be in combo.cpp
Since dpll_enabled_list is currently only used in combo.cpp, we should
be cleaning up the list here as well. Also, previously, we were cleaning
it up once when the entire program was exiting, but we should be doing
this when we are done waiting for all the combo phy timers.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>