Resolve PLL Programming and Div by 0 (#8)
* DKL: Fix a bug with multiple DKL phys not being seen

Each Dekel PHY is addressed through a 4KB aperture. Each PHY has more than
4KB of register space, so a separate index is programmed in HIP_INDEX_REG0
or HIP_INDEX_REG1, based on the port number, to set the upper 2 address
bits that point the 4KB window into the full PHY register space.
So, basically, if we read the value of the PLL DIV0 register and it is 0,
then that means that this aperture may need to be shifted. Once we do that,
we should re-read the value of the DIV0 register once more and this time it
should be valid. During resetting of the DKL phys, it is possible that the register index
moves back to another window and as such programming of the phy
registers has no impact. So before each programming, ensure that we are
in the right window.

* vbltest: Provide a cmd param to get Vsync interrupts on a per pipe basis

Earlier, we only used to get the vsyncs interrupts for the primary pipe.
This means, we couldn't synchronize other pipes because we didn't know
when their vsync interrupts were happening. With this change, this
program takes an extra parameter in the form of a 0 based pipe for which
we need to get the vsyncs. So the usage of vbltest is:

vbltest <number of vsyncs to get timestamp for> <0 based pipe to get for>

* Combo: Fix a divide by 0 error

According to the Bspec, we have to get the qdiv_mode first from cfgcr1
register's bit 9 and only if it is enabled, should we read qdiv_ratio
from cfgcr1's bit 17:10. Otherwise, it should be 1. We were incorrectly
not reading this bit 9 and that would cause us to read bits 17:10 even
when they were 0 which would end up causing a divide by 0 error. This
change addresses this problem.
---------

Signed-off-by: Jubilee Steinbrink <jubilee.steinbrink@intel.com>
Signed-off-by: Bill Mahoney <bill.mahoney@intel.com>
Signed-off-by: Eshe N Pickett <eshe.n.pickett@intel.com>
Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Co-authored-by: Jubilee Steinbrink <jubilee.steinbrink@intel.com>
Co-authored-by: Bill Mahoney <bill.mahoney@intel.com>
Co-authored-by: Brian McGinn <brian.mcginn@intel.com>
Co-authored-by: Satyeshwar Singh <satyeshwar.singh@intel.com>