Combo fix (#9)
* vsync: Rename check to wait_until

The check_if_done has been renamed to wait_until_done because it was
technically waiting until all the timers are done. Also, made the code a
bit cleaner by deleting the timer inside the wait_until_done rather than
in synchronize_phys.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>

* combo: Moved cleanup_list to be in combo.cpp

Since dpll_enabled_list is currently only used in combo.cpp, we should
be cleaning up the list here as well. Also, previously, we were cleaning
it up once when the entire program was exiting, but we should be doing
this when we are done waiting for all the combo phy timers.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>

* vsync: Add register addresses for debugging purposes

When printing the register names and values, it is also helpful if the
register addresses are printed.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>

---------

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Signed-off-by: Eshe N Pickett <eshe.n.pickett@intel.com>
Co-authored-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Signed-off-by: Eshe N Pickett <eshe.n.pickett@intel.com>
Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Co-authored-by: Jubilee Steinbrink <jubilee.steinbrink@intel.com>
Co-authored-by: Bill Mahoney <bill.mahoney@intel.com>
Co-authored-by: Brian McGinn <brian.mcginn@intel.com>
Co-authored-by: Satyeshwar Singh <satyeshwar.singh@intel.com>