vsync: Reorg'd into classes for Phys
With this commit, we have completely reorganized the code such that it
is more object oriented and using inheritance. Also, at initialization,
we will find out all the PHYs that are currently present and if one of
the PLLs is being used for multiple CRTCs. In such a scenario, the code
would report an error at init time.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>