lib: Allow user to synchronize individual pipes
By default, synchronize_vsyncs only took one parameter which was
time_diff and synchronized all pipes. However, there are circumstances
when the caller wants to provide one pipe at a time because other pipes
have to be synchronized to a different time delta. With this change, we
introduce a second optional parameter through which the caller can
provide the individual 0 based pipe that they want to synchronize. Note
that they must know which pipe number that they need to synchronize. If
they don't specify this parameter or specify it to ALL_PIPES, old
functionality of synchronizing all the pipes will continue.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>